added seed variable to database

+++ migrations/20160824205535_create_team_game_stats.js
@@ -7,7 +7,7 @@ exports.up = function(knex, Promise) {
     t.integer('pointDiff');
     t.integer('pointsFor');
     t.integer('pointsAgainst');
-    t.integer('seed');
+    t.decimal('seed');
     t.timestamps();
   })
 };
